Increased granularity of stage growth time generation from seconds to ticks; config files still use seconds as their unit of time

+++ src/me/kosinkadink/performantplants/util/TimeHelper.java
@@ -2,11 +2,11 @@ package me.kosinkadink.performantplants.util;
 
 public class TimeHelper {
 
-    public static int secondsToTicks(int seconds) {
+    public static long secondsToTicks(int seconds) {
         return seconds * 20;
     }
 
-    public static int minutesToTicks(int minutes) {
+    public static long minutesToTicks(int minutes) {
         return secondsToTicks(minutes * 60);
     }
